# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_orbit_internal_resputil.Response-any:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_orbit_internal_resputil.Response-internal_handler_PutUserInProjectResp:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/internal_handler.PutUserInProjectResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_orbit_internal_resputil.Response-string:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data:
        type: string
      msg:
        type: string
    type: object
...
  internal_handler.AdjustUserExtraBalanceReq:
    properties:
      delta:
        type: integer
      reason:
        type: string
    required:
    - delta
    type: object
...
  internal_handler.SharedUserReq:
    properties:
      datasetID:
        type: integer
      userIDs:
        items:
          type: integer
        type: array
    required:
    - datasetID
    - userIDs
    type: object
...
  internal_handler.UpdateRoleReq:
    properties:
      role:
        $ref: '#/definitions/github_com_raids-lab_orbit_dao_model.Role'
    required:
    - role
    type: object
...
  internal_handler.UserPutUserInProjectReq:
    properties:
      accessmode:
        type: string
      quota:
        $ref: '#/definitions/datatypes.JSONType-github_com_raids-lab_orbit_dao_model_QueueQuota'
      role:
        type: string
    type: object
...
  internal_handler.cancelsharedUserReq:
    properties:
      datasetID:
        type: integer
      userID:
        type: integer
    required:
    - datasetID
    - userID
    type: object
...
info:
  contact: {}
  description: This is the API server for Orbit, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Orbit API
  version: 1.0.0
paths:
...
  /v1/accounts/{aid}/users:
    get:
      consumes:
      - application/json
      description: Get list of users in account (requires user to be in account)
      parameters:
      - description: aid
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User account entries
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "403":
          description: Forbidden - user not in account
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Get users in account
      tags:
      - Project
  /v1/accounts/{aid}/users/{uid}:
    delete:
      consumes:
      - application/json
      description: Account admin removes user from account
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: integer
      - description: aid
        in: path
        name: aid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns removed username and account name
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "403":
          description: Forbidden - not account admin
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Remove user from account (account admin)
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Account admin adds user to account
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: integer
      - description: aid
        in: path
        name: aid
        required: true
        type: integer
      - description: Role and access mode
        in: body
        name: req
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: Returns added username and account name
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "403":
          description: Forbidden - not account admin
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Add user to account (account admin)
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Account admin batch updates user-account relationship (partial
        update)
      parameters:
      - description: aid
        in: path
        name: aid
        required: true
        type: integer
      - description: uid
        in: path
        name: uid
        required: true
        type: integer
      - description: Update data
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UserPutUserInProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: Update result
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-internal_handler_PutUserInProjectResp'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "403":
          description: Forbidden - not account admin
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Batch update user-account relationship (account admin)
      tags:
      - Project
...
  /v1/admin/dataset/cancelshare/user:
    post:
      consumes:
      - application/json
      description: 管理员取消数据集共享用户
      parameters:
      - description: 共享数据集用户
        in: body
        name: Req
        required: true
        schema:
          $ref: '#/definitions/internal_handler.cancelsharedUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 管理员取消数据集共享用户
      tags:
      - Dataset
...
  /v1/admin/dataset/share/user:
    post:
      consumes:
      - application/json
      description: 管理员对用户分享数据集
      parameters:
      - description: 共享数据集用户
        in: body
        name: userReq
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SharedUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 管理员对用户分享数据集
      tags:
      - Dataset
...
  /v1/admin/users:
    get:
      consumes:
      - application/json
      description: 列出用户信息（包含私人配额）
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户信息
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 列出用户信息
      tags:
      - User
  /v1/admin/users/{name}:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: username
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - User
  /v1/admin/users/{name}/attributes:
    put:
      consumes:
      - application/json
      description: 管理员更新指定用户的属性
      parameters:
      - description: username
        in: path
        name: name
        required: true
        type: string
      - description: 用户属性
        in: body
        name: attributes
        required: true
        schema:
          $ref: '#/definitions/github_com_raids-lab_orbit_dao_model.UserAttribute'
      produces:
      - application/json
      responses:
        "200":
          description: 用户属性更新成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 管理员更新用户属性
      tags:
      - User
  /v1/admin/users/{name}/billing/extra-balance:
    post:
      consumes:
      - application/json
      description: 管理员按增量调整用户 extraBalance（可正可负），使用行锁避免并发覆盖
      parameters:
      - description: username
        in: path
        name: name
        required: true
        type: string
      - description: 增量与原因
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AdjustUserExtraBalanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 调整成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-internal_handler_AdjustUserExtraBalanceResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 调整用户额外点数
      tags:
      - User
  /v1/admin/users/{name}/role:
    put:
      consumes:
      - application/json
      description: 更新角色
      parameters:
      - in: path
        name: name
        required: true
        type: string
      - description: role
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新角色成功
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 更新角色
      tags:
      - User
  /v1/admin/users/baseinfo:
    get:
      consumes:
      - application/json
      description: 列出用户信息，姓名，昵称，用户空间
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户信息
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: 其他错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 列出用户基本信息
      tags:
      - User
...
  /v1/dataset/cancelshare/user:
    post:
      consumes:
      - application/json
      description: 普通用户取消数据集共享
      parameters:
      - description: 共享数据集用户
        in: body
        name: Req
        required: true
        schema:
          $ref: '#/definitions/internal_handler.cancelsharedUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回值描述
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 普通用户取消数据集共享用户
      tags:
      - Dataset
...
  /v1/images/user:
    get:
      consumes:
      - application/json
      description: 获取未被分享该镜像的用户
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取未被分享该镜像的用户（支持名称模糊搜索）
      tags:
      - ImagePack
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"